TX details screen (#788)
Add more sections to details screen

* Remove Kotlin Synthetics
* Pull to refresh not crashing anymore
* Type and status for custom and modify settings
* Show txData section only for custom transfer
* Move to client gateway data model
* Remove failure handler. safeLaunch() already logs errors.
* Added TODO to use executor, when available
* Remove TxDataView from Transfer and SettingsChange. Format xml files
* Left/Right -> Start/End + animateLayoutChanges
* Data classes renamed to *Dto
* Removed Domain prefix

+++ data/src/main/java/io/gnosis/data/models/Transaction.kt
@@ -19,7 +19,7 @@ sealed class Transaction {
         override val confirmations: Int?,
         val nonce: BigInteger?,
         val address: Solidity.Address,
-        val dataSize: Long,
+        val dataSize: Int,
         val date: Date?,
         val value: BigInteger
     ) : Transaction()
